Move shared mixins to mixins.py
Include CustomIcon class as it is only used by NavMenuItemMediaMixin
